Move server{_test}.go to new package api/
