+++ backend/api/testdata/server/golden/the-new-yorker.json
